requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Caching==2.1.0
Flask-WTF==1.1.1
WTForms==3.0.1

# Database
SQLAlchemy==1.4.49

# Utilities
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0

# Development
pytest==7.4.0
pytest-flask==1.2.0

# Production
gunicorn==21.2.0